+++ packages/chrome/public/devtools.js
@@ -1,18 +1,17 @@
-const checkForThemeUi = () => {
+const checkForThemeUI = () => {
   return new Promise((resolve, reject) => {
     window.chrome.devtools.inspectedWindow.eval(
       `window.__THEME_UI__.theme`,
-      (result, err) => (err ? reject(err) : resolve(result))
+      (result, err) => (err ? resolve(false) : resolve(true))
     )
   })
 }
 
-chrome.devtools.panels.elements.createSidebarPane('Theme UI', sidebar => {
-  checkForThemeUi()
-    .then(() => {
+checkForThemeUI()
+  .then((hasThemeUI) => {
+    if (!hasThemeUI) return
+    chrome.devtools.panels.elements.createSidebarPane('Theme UI', sidebar => {
       sidebar.setPage('panel.html')
     })
-    .catch(() => {
-      sidebar.setPage('disabled.html')
-    })
-})
+  })
+